some refactor, mock filters
