+++ src/mock/destinations.js
@@ -1,5 +1,5 @@
-import {COUNTRIES, IMAGE_REFERENCE, MAX_IMAGE_ID, POINT_DESCRIPTIONS} from '../const';
-import {getRandomInteger} from '../util';
+import {COUNTRIES, IMAGE_REFERENCE, MAX_IMAGE_ID, POINT_DESCRIPTIONS} from '../utils/const';
+import {getRandomInteger} from '../utils/util';
 
 const getRandomDescription = () => POINT_DESCRIPTIONS[getRandomInteger(POINT_DESCRIPTIONS.length - 1)];
 const getRandomCountry = () => COUNTRIES[getRandomInteger(COUNTRIES.length - 1)];
